+++ packages/@overeng/genie/nix/build.nix
@@ -18,7 +18,7 @@ let
     # Patches referenced in pnpm-workspace.yaml (shared across all workspaces)
     patchesDir = "packages/@overeng/utils/patches";
     # Managed by `dt nix:hash:genie` — do not edit manually.
-    pnpmDepsHash = "sha256-0OSGcThI7Q9YhYLhNmTpobn7/2IYnDg0ejmisZcaoSQ=";
+    pnpmDepsHash = "sha256-p9ZPQ8Gtnpaitmpx6YR8rPHrUy3oK20ojx2a6BZeb+E=";
     lockfileHash = "sha256-A8axH5fMO9VA3l9Sh68TdphefBGhcovgUwwy1tentsI=";
     packageJsonDepsHash = "sha256-tMch41qH+GilQSbIGitAHjKtPH2tb4h9uSwDW1peQDc=";
     inherit gitRev commitTs dirty;