+++ packages/@overeng/megarepo/nix/build.nix
@@ -16,7 +16,7 @@ let
     # Patches are in packages/@overeng/utils/patches/ (referenced by pnpm-lock.yaml)
     patchesDir = "packages/@overeng/utils/patches";
     # Managed by `dt nix:hash:megarepo` — do not edit manually.
-    pnpmDepsHash = "sha256-Dk+XgTDM73qoHBAL90T2UNuEcUHb2iObo9FwC2jTfTE=";
+    pnpmDepsHash = "sha256-P9QSZ1460ifeXKw2veuqfHJMyxZfXBy/tmvXHedFvlI=";
     lockfileHash = "sha256-s3mK+bc5FIB2RI7ZOAUFIeQU7U3tL7tytpoEFfxkCy4=";
     packageJsonDepsHash = "sha256-dhrqTHhoUUk753oIk61dMTbxJ1ivA9UYEu6h9jM51qA=";
     smokeTestArgs = [ "--help" ];